Issue #22 handle case of generation 99999 going to 00000

+++ als-computation/src/main/java/com/cloudera/oryx/als/computation/iterate/row/RowStep.java
@@ -57,7 +57,7 @@ public final class RowStep extends IterationStep {
 
     JobStepConfig config = getConfig();
     String instanceDir = config.getInstanceDir();
-    long generationID = config.getGenerationID();
+    int generationID = config.getGenerationID();
 
     if (store.exists(Namespaces.getInstanceGenerationPrefix(instanceDir, generationID) + "X/", false)) {
       // Actually, looks like whole computation of X/Y finished -- just proceed
