+++ als-computation/src/main/java/com/cloudera/oryx/als/computation/merge/MergeNewOldStep.java
@@ -49,8 +49,8 @@ public final class MergeNewOldStep extends ALSJobStep {
     JobStepConfig jobConfig = getConfig();
 
     String instanceDir = jobConfig.getInstanceDir();
-    long generationID = jobConfig.getGenerationID();
-    long lastGenerationID = jobConfig.getLastGenerationID();
+    int generationID = jobConfig.getGenerationID();
+    int lastGenerationID = jobConfig.getLastGenerationID();
 
     String outputKey = Namespaces.getInstanceGenerationPrefix(instanceDir, generationID) + "input/";
     if (!validOutputPath(outputKey)) {
